Moved controlelrs base parts
