+++ src/components/Controllers/Metadata/Panels/Report.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import ReportPanel from '../../../Parts/Panels/Report'
+import ReportPanel from '../../Base/Panels/Report'
 import { useStore } from '../store'
 
 export default function Report() {
